form types and util enough to implement GetNumItems

+++ obse64/GameScript.h
@@ -77,7 +77,7 @@ typedef bool (* ExecuteFunction)(
 	TESObjectREFR * containingObj,
 	Script * script,
 	ScriptLocals * locals,
-	float * result,
+	double * result,
 	u32 * opcodeOffsetPtr);
 
 typedef bool (* ParseFunction)(
@@ -131,7 +131,7 @@ typedef bool (* _ExtractArgs)(
 	...);
 extern RelocAddr <_ExtractArgs> ExtractArgs;
 
-#define COMMAND_ARGS		const ParamInfo * paramInfo, const char * scriptData, TESObjectREFR * thisObj, TESObjectREFR * containingObj, Script * script, ScriptLocals * locals, float * result, u32 * opcodeOffsetPtr
+#define COMMAND_ARGS		const ParamInfo * paramInfo, const char * scriptData, TESObjectREFR * thisObj, TESObjectREFR * containingObj, Script * script, ScriptLocals * locals, double * result, u32 * opcodeOffsetPtr
 #define PASS_COMMAND_ARGS	paramInfo, scriptData, thisObj, containingObj, script, locals, result, opcodeOffsetPtr
 #define EXTRACT_ARGS		paramInfo, scriptData, opcodeOffsetPtr, thisObj, containingObj, script, locals
 #define COMMAND_ARGS_EVAL	TESObjectREFR * thisObj, void * arg1, void * arg2, double * result
